+++ src/gpu/lcd/sprites/sprite_info.rs
@@ -1,6 +1,6 @@
 pub struct SpriteInfo {
     pub attributes: u8,
     pub tile_location: u16,
-    pub x_position: u8,
-    pub y_position: u8
+    pub x_position: i32,
+    pub y_position: i32
 }
